requirements.txt
boto3==1.34.0
pandas==2.1.4
numpy==1.24.3
pyspark==3.5.0
awswrangler==3.5.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pyarrow==14.0.2
great-expectations==0.18.8
aws-cdk-lib==2.117.0
constructs>=10.0.0,<11.0.0
pytest==7.4.3
moto==4.2.14
boto3-stubs[essential]==1.34.0
